разносит фунции из helpers
