+++ src/main.js
@@ -8,7 +8,7 @@ import ShowMoreButton from "./view/show-more.js";
 import FilmsExtra from "./view/films-extra.js";
 import PopUpFilmCard from "./view/popup.js";
 import {generateCard} from './mock/cards.js';
-import {RenderPosition, renderElement} from './helpers.js';
+import {RenderPosition, renderElement, remove} from './helpers/render.js';
 
 const CARDS_IN_ROW = 5;
 const NUMBER_OF_EXTRA_CONTAINERS = 2;
@@ -26,7 +26,6 @@ const siteBody = document.querySelector(`body`);
 const renderFilm = (filmListElement, film) => {
   const filmComponent = new FilmCard(film);
   const filmPopup = new PopUpFilmCard(film);
-  // const taskEditComponent = new TaskEditView(task);
 
   const closePopup = () => {
     filmPopup.getElement().remove();
@@ -42,8 +41,7 @@ const renderFilm = (filmListElement, film) => {
   };
 
   const openPopup = () => {
-    // evt.preventDefault();
-    renderElement(siteBody, filmPopup.getElement(), RenderPosition.BEFOREEND);
+    renderElement(siteBody, filmPopup, RenderPosition.BEFOREEND);
     const closePopupBtn = filmPopup.getElement().querySelector(`.film-details__close-btn`);
     siteBody.classList.add(`hide-overflow`);
     closePopupBtn.addEventListener(`click`, closePopup);
@@ -52,24 +50,23 @@ const renderFilm = (filmListElement, film) => {
 
   filmComponent.showPopupClickHandler(() => {
     openPopup();
-    document.removeEventListener(`keydown`, onEscKeyDown);
   });
 
-  renderElement(filmListElement, filmComponent.getElement(), RenderPosition.BEFOREEND);
+  renderElement(filmListElement, filmComponent, RenderPosition.BEFOREEND);
 };
 
 const favorCount = generatedCards.filter((card) => card.isFavorite).length;
 const watchedCount = generatedCards.filter((card) => card.isWatched).length;
 const isInWatchListCount = generatedCards.filter((card) => card.isInWatchList).length;
 
-renderElement(siteHeaderElement, new Profile().getElement(), RenderPosition.BEFOREEND);
-renderElement(siteMainElement, new Navigation(isInWatchListCount, watchedCount, favorCount).getElement(), RenderPosition.BEFOREEND);
-renderElement(siteMainElement, new Sort().getElement(), RenderPosition.BEFOREEND);
-renderElement(siteMainElement, new Films().getElement(), RenderPosition.BEFOREEND);
+renderElement(siteHeaderElement, new Profile(), RenderPosition.BEFOREEND);
+renderElement(siteMainElement, new Navigation(isInWatchListCount, watchedCount, favorCount), RenderPosition.BEFOREEND);
+renderElement(siteMainElement, new Sort(), RenderPosition.BEFOREEND);
+renderElement(siteMainElement, new Films(), RenderPosition.BEFOREEND);
 
 const siteFilms = siteMainElement.querySelector(`.films`);
 
-renderElement(siteFilms, new FilmsList().getElement(), RenderPosition.BEFOREEND);
+renderElement(siteFilms, new FilmsList(), RenderPosition.BEFOREEND);
 
 const siteFilmsList = siteFilms.querySelector(`.films-list`);
 const siteFilmsListContainer = siteFilmsList.querySelector(`.films-list__container`);
@@ -81,7 +78,7 @@ for (let i = 0; i < Math.min(generatedCards.length, CARDS_IN_ROW); i++) {
 if (generatedCards.length > CARDS_IN_ROW) {
   const loadMoreButtonComponent = new ShowMoreButton();
 
-  renderElement(siteFilms, loadMoreButtonComponent.getElement(), RenderPosition.BEFOREEND);
+  renderElement(siteFilms, loadMoreButtonComponent, RenderPosition.BEFOREEND);
 
   loadMoreButtonComponent.loadMoreClickHandler(() => {
     // evt.preventDefault();
@@ -92,8 +89,7 @@ if (generatedCards.length > CARDS_IN_ROW) {
     showingFilmCardsCount += CARDS_IN_ROW;
 
     if (showingFilmCardsCount > generatedCards.length) {
-      loadMoreButtonComponent.getElement().remove();
-      loadMoreButtonComponent.removeElement();
+      remove(loadMoreButtonComponent);
     }
   });
 }
@@ -101,7 +97,7 @@ if (generatedCards.length > CARDS_IN_ROW) {
 let headers = [`Top rated`, `Most recommended`];
 
 for (let i = 0; i < NUMBER_OF_EXTRA_CONTAINERS; i++) {
-  renderElement(siteFilms, new FilmsExtra(headers[i]).getElement(), RenderPosition.BEFOREEND);
+  renderElement(siteFilms, new FilmsExtra(headers[i]), RenderPosition.BEFOREEND);
 }
 
 // const filmsExtraContainers = siteMainElement.querySelectorAll(`.films-list--extra .films-list__container`);
@@ -115,11 +111,11 @@ const topRatedFilms = [...generatedCards]
   .slice(-TOP_RATED_FILMS);
 
 topRatedFilms
-  .forEach((film) => renderElement(topRated, new FilmCard(film).getElement(), RenderPosition.BEFOREEND));
+  .forEach((film) => renderElement(topRated, new FilmCard(film), RenderPosition.BEFOREEND));
 
 generatedCards
   .slice(0, MOST_COMMENTED_FILMS)
-  .forEach((film) => renderElement(mostCommented, new FilmCard(film).getElement(), RenderPosition.BEFOREEND));
+  .forEach((film) => renderElement(mostCommented, new FilmCard(film), RenderPosition.BEFOREEND));
 
 
 // for (let i = 0; i < NUMBER_OF_EXTRA_CONTAINERS; i++) {
